+++ spec/default_spec.rb
@@ -1,19 +1,12 @@
-# encoding: UTF-8
 describe 'resharper::default' do
-  describe 'version 10' do
-    before :each do
-      ENV['WINDIR'] = 'C:\Windows'
-    end
+  describe 'version 2016.2.2' do
     let(:chef_run) do
-      ChefSpec::ServerRunner.new do |node|
-        node.set['resharper']['version'] = '10'
-        node.set['visualstudio']['source'] = 'http://blah.com'
-      end.converge(described_recipe)
+      ChefSpec::SoloRunner.new(platform: 'windows', version: '2008R2').converge(described_recipe)
     end
     it 'download finished notifies install executable' do
-      resource = chef_run.remote_file('download_resharper_10')
+      resource = chef_run.remote_file('download_resharper_2016.2.2')
       expect(resource).to notify(
-        'execute[install_resharper_10]'
+        'execute[install_resharper_2016.2.2]'
       ).to(:run).immediately
     end
   end