+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # CHANGELOG for resharper
 
 This file is used to list changes made in each version of resharper.
-## 0.1.0:
-Initial version.
+
+## 2.0
+- Added support for installing ReSharper 2016+, which uses a non-MSI installer.
+
+## 1.0
+- Supports installing ReSharper 7.x
+
 - - - 
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.